+++ src/main/java/com/epam/reportportal/cucumber/ScenarioContext.java
@@ -19,9 +19,9 @@ package com.epam.reportportal.cucumber;
 import io.cucumber.plugin.event.Node;
 import io.cucumber.plugin.event.TestCase;
 import io.reactivex.Maybe;
+import jakarta.annotation.Nonnull;
+import jakarta.annotation.Nullable;
 
-import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
 import java.net.URI;
 import java.util.Optional;