+++ src/test/java/com/epam/reportportal/cucumber/ScenarioOutlineStepReporterTest.java
@@ -42,8 +42,8 @@ import java.util.stream.Stream;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.*;
-import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.any;
 
 /**
  * @author <a href="mailto:ihar_kahadouski@epam.com">Ihar Kahadouski</a>
@@ -117,10 +117,12 @@ public class ScenarioOutlineStepReporterTest {
 
 		List<String> items = testCaptor.getAllValues().stream().map(StartTestItemRQ::getName).collect(Collectors.toList());
 
-		assertThat(items, hasItems(
-				"Scenario Outline: Test with the parameter \"first\"",
-				"Scenario Outline: Test with the parameter \"second\"",
-				"Scenario Outline: Test with the parameter \"third\""
-		));
+		assertThat(
+				items, hasItems(
+						"Scenario Outline: Test with the parameter \"first\"",
+						"Scenario Outline: Test with the parameter \"second\"",
+						"Scenario Outline: Test with the parameter \"third\""
+				)
+		);
 	}
 }